hotfix menu update

+++ src/Infrastructure/Integrations/WelcomeGroup/DataTransferObjects/Food/EditFoodResponseData.php
@@ -42,7 +42,7 @@ final class EditFoodResponseData extends ResponseData
             $this->description,
             $this->weight,
             $this->caloricity,
-            $this->price * 100,
+            (int)($this->price * 100),
         );
     }
 }
